pyproject.toml
[project]
name = "numerai-batch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "catboost==1.2",
    "cloudpickle==2.2.1",
    "google-cloud-batch>=0.17.35",
    "google-cloud-storage>=3.1.0",
    "ipykernel>=6.29.5",
    "lightgbm==4.0.0",
    "matplotlib==3.7.1",
    "numerapi==2.20.6",
    "numpy==1.23.5",
    "pandera>=0.24.0",
    "plotly==5.14.1",
    "polars==0.18.15",
    "pyarrow==11.0.0",
    "python-dotenv>=1.1.0",
    "scikit-learn==1.2.2",
    "xgboost==1.7.5",
]

[tool.ruff]
target-version = "py311"
include = ["*.py"]
exclude = ["src/alembic/versions/*.py"]
line-length = 119

[tool.ruff.lint]
select = [
  "C90",
  "D",
  "E",
  "F",
  "I",
  "W",
]
ignore = [
  "D100",
  "E501",
  "D400",
  "D415",
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["musubi_restapi"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
split-on-trailing-comma = true

[dependency-groups]
dev = [
    "ruff>=0.11.6",
    "ty>=0.0.1a1",
]
